update seed_agent_no_update to use epsilon value
epsilon value represents the mutation probability for the prototoype
that will be used to seed the agent.  the update to the
seed_agent_no_update asserts that there is a prototype, the prototype is
a list, and if epsilon is zero, we set the prototype to the state.  if
epsilon is greater than zero, we use the newly created mutate function
to mutate the prototype, and set the state as that