_call_lens no longer a hidden, renamed param name, and added docstring
call lens is now a public method.  This will allow child classes to
inhered the method to use.

added doc string

renamed param from lens_in_file to Lens_in_file_dir to signify that it
needs the directory and .in file to be passed

also deleted commented out lines in function